Using float instead of double

+++ c/estimates.c
@@ -2,13 +2,13 @@
 #include <math.h>
 #include "parameters.h"
 
-double** s(size_t n, double* p) {
-    double** matrix = (double**)malloc((n+1) * sizeof(double*));
-    matrix[0] = (double*)malloc(sizeof(double));
+float** s(size_t n, float* p) {
+    float** matrix = (float**)malloc((n+1) * sizeof(float*));
+    matrix[0] = (float*)malloc(sizeof(float));
     matrix[0][0] = 1;
     size_t k, m = 0;
     do {
-        matrix[m+1] = (double*)malloc((m+2) * sizeof(double));
+        matrix[m+1] = (float*)malloc((m+2) * sizeof(float));
         matrix[m+1][0] = matrix[m][0] * (1 - p[m]);
         matrix[m+1][m+1] = matrix[m][m] * p[m];
         k = 0;
@@ -20,9 +20,9 @@ double** s(size_t n, double* p) {
     return matrix;
 }
 
-double R(size_t n, size_t r, double* p) {
-    double** s_matrix = s(n, p);
-    double result = 0;
+float R(size_t n, size_t r, float* p) {
+    float** s_matrix = s(n, p);
+    float result = 0;
     do {
         result += s_matrix[n][r];
     } while(r++ < n);
@@ -34,13 +34,13 @@ double R(size_t n, size_t r, double* p) {
     return result;
 }
 
-double Q(double R, double rho, size_t n, double* a, double* t) {
-    long double exponent = 0.0;
+float Q(float R, float rho, size_t n, float* a, float* t) {
+    float exponent = 0.0;
     size_t i = 0;
     do {
         exponent += (1 - a[i]) * t[i];
     } while (++i < n);
-    double result = R * exp(rho * exponent);
+    float result = R * exp(rho * exponent);
     while (n-- > 0) {
         result *= a[n];
     }
